refactor: move replicate ai
