+++ packages/llm/replicate/src/index.edge-light.ts
@@ -1,0 +1,5 @@
+export class ReplicateLLM {
+  constructor() {
+    throw new Error("ReplicateLLM is not supported in vercel edge runtime");
+  }
+}
