+++ packages/llm/replicate/src/index.ts
@@ -1,0 +1,8 @@
+export {
+  ALL_AVAILABLE_REPLICATE_MODELS,
+  DeuceChatStrategy,
+  LlamaDeuce,
+  ReplicateChatStrategy,
+  ReplicateLLM,
+  ReplicateSession,
+} from "./llm";
